adds feat manager that also gets list of feat names
